service_analysis/requirements.txt
# Service Analysis - Core Dependencies

# Web Framework
fastapi==0.111.0
uvicorn==0.30.1
pydantic

# Google Drive API
google-api-python-client==2.137.0
google_auth_oauthlib
python-dotenv==1.0.1

# AI/ML - BLIP Image Captioning
torch==1.12.1+cu113
torchvision==0.13.1+cu113
-f https://download.pytorch.org/whl/torch_stable.html
transformers==4.30.2
huggingface_hub==0.14.1

# NLP - Text Processing & ASO Keywords
sentence-transformers==2.2.2
nltk==3.8.1

# Core Libraries
numpy<2.0
opencv-python
Pillow==10.0.1
